requirements.txt
# Goodbooks-10k рекомендательный сервис
# Python 3.8+

pandas>=1.3.0
numpy>=1.20.0,<2.0.0
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-surprise>=1.1.4
matplotlib>=3.5.0
tabulate>=0.9.0
